[refactoring - bigstars] delete nova patterns + uint Planet type + uint Star type + smaller bigstars + rarify some nova patterns

+++ shaders/fragment/planets/common.glsl
@@ -3,7 +3,7 @@
 
 struct Planet
 {
-  float type;
+  uint type;
   vec2 center;
   float rotation;
   float radius;
